Being a dummy... moved some real code out of the test folder
